Provide secretkey to validate user name and it's credentials

+++ app.js
@@ -10,6 +10,9 @@ var bodyParser = require('body-parser');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true })); 
 
+var apiUser = 'root';
+var apiKey = 'iloveberries';
+
 var check;
 
 //db.serialize(function() {
@@ -19,14 +22,14 @@ var check;
 app.get('/api/v1/users/', function(req, res) {
   console.log('landed in users\n')
   db.all("SELECT username FROM users", function(err, users) {
-    res.status(200).json({users});
+    return res.status(200).json({users});
   });
 })
 
 app.get('/api/v1/users/:userName', function(req, res) {
   console.log("SELECT username from users where userName = '" + req.params.userName + "'")
   db.all("SELECT * from users where userName = '" + req.params.userName + "'", function(err, users){
-    res.status(200).json({users})
+    return res.status(200).json({users})
   });
 })
 
@@ -35,15 +38,26 @@ app.post('/api/v1/users/', function(req, res) {
   var stmt = db.prepare("INSERT into users VALUES (?, ?, ?)", req.body.username, req.body.email, hash);
   stmt.run();
   stmt.finalize();
-  res.status(200).send('Success - user added');
+  return res.status(200).send('Success - user added');
 })
 
 app.delete('/api/v1/users/:userName', function(req, res) {
   console.log(req.body);
+  if (!req.headers.authorization) {
+    return res.json({ error: 'No credentials sent!' });
+  }
+
+  var encoded = req.headers.authorization.split(' ')[1];
+  var decoded = new Buffer(encoded, 'base64').toString('utf8');
+
+  if (decoded.split(':')[0] != apiUser || decoded.split(':')[1] != apiKey) {
+    return res.json({ error: 'Wrong credentials' });
+  }
+
   var stmt = db.prepare("DELETE from users where username = '" + req.params.userName + "'")
   stmt.run();
   stmt.finalize();
-  res.status(200).send('landed in users delete\n')
+  return res.status(200).send('landed in users delete\n')
 })
 
 app.listen(8000, function() {
